# docs/swagger.yaml
basePath: /
definitions:
...
  response.Error:
    properties:
      code:
        example: INVALID_INPUT
        type: string
      error:
        example: Invalid request
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  response.ValidationError:
    properties:
      error:
        example: Validation failed
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
      success:
        example: false
        type: boolean
    type: object
  review.CreateReviewRequest:
    description: Create review request
    properties:
      comment:
        example: Great work, very professional!
        type: string
      job_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      rating:
        example: 5
        type: integer
    type: object
  review.RatingSummaryResponse:
    description: Rating summary
    properties:
      average_rating:
        example: 4.7
        type: number
      total_reviews:
        example: 12
        type: integer
    type: object
  review.ReviewListResponse:
    description: Paginated review list
    properties:
      has_more:
        example: false
        type: boolean
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
      reviews:
        items:
          $ref: '#/definitions/review.ReviewResponse'
        type: array
      total:
        example: 12
        type: integer
    type: object
  review.ReviewResponse:
    description: Review information
    properties:
      comment:
        example: Great work, very professional!
        type: string
      created_at:
        example: "2025-01-15T14:30:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      job_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      rating:
        example: 5
        type: integer
      reviewee_id:
        example: 550e8400-e29b-41d4-a716-446655440003
        type: string
      reviewer_id:
        example: 550e8400-e29b-41d4-a716-446655440002
        type: string
      type:
        example: client_to_handyman
        type: string
    type: object
...
  wallet.AdminTopUpRequest:
    description: Admin top-up request
    properties:
      amount:
        example: 100
        type: integer
      description:
        example: 'Package purchase: Starter 100 PLN'
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
    type: object
  wallet.TransactionListResponse:
    description: Paginated transaction list
    properties:
      has_more:
        example: true
        type: boolean
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
      total:
        example: 25
        type: integer
      transactions:
        items:
          $ref: '#/definitions/wallet.TransactionResponse'
        type: array
    type: object
  wallet.TransactionResponse:
    description: Wallet transaction record
    properties:
      amount:
        example: 22
        type: integer
      balance_after:
        example: 65
        type: integer
      created_at:
        example: "2025-01-15T14:30:00Z"
        type: string
      description:
        example: 'Lead accepted: Hydraulik - Krowodrza'
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      reason:
        example: lead_accepted
        type: string
      reference_id:
        example: 550e8400-e29b-41d4-a716-446655440002
        type: string
      type:
        example: debit
        type: string
    type: object
  wallet.WalletResponse:
    description: Wallet balance information
    properties:
      balance:
        example: 87
        type: integer
      created_at:
        example: "2025-01-10T12:00:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      updated_at:
        example: "2025-01-15T14:30:00Z"
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@fixapp.com
    name: API Support
  description: FixApp backend API server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: FixApp API
  version: "1.0"
paths:
...
  /admin/wallet/{user_id}:
    get:
      consumes:
      - application/json
      description: Returns a user's wallet balance (admin only)
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.WalletResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Get user wallet (admin)
      tags:
      - admin
  /admin/wallet/top-up:
    post:
      consumes:
      - application/json
      description: Adds credits to a user's wallet (admin only)
      parameters:
      - description: Top-up details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/wallet.AdminTopUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.WalletResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ValidationError'
      security:
      - BearerAuth: []
      summary: Top up user wallet (admin)
      tags:
      - admin
...
  /jobs/{id}/reviews:
    get:
      consumes:
      - application/json
      description: Returns all reviews for a specific completed job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/review.ReviewResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: List reviews for a job
      tags:
      - reviews
...
  /reviews:
    post:
      consumes:
      - application/json
      description: Submit a review for a completed job. Clients review handymen, handymen
        review clients.
      parameters:
      - description: Review data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/review.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/review.ReviewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Create a review
      tags:
      - reviews
...
  /users/{id}/rating:
    get:
      consumes:
      - application/json
      description: Returns average rating and total review count for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/review.RatingSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
      summary: Get rating summary for a user
      tags:
      - reviews
  /users/{id}/reviews:
    get:
      consumes:
      - application/json
      description: Returns paginated reviews received by a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Limit (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/review.ReviewListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
      summary: List reviews for a user
      tags:
      - reviews
  /wallet:
    get:
      consumes:
      - application/json
      description: Returns the authenticated user's wallet balance
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.WalletResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Get my wallet
      tags:
      - wallet
  /wallet/transactions:
    get:
      consumes:
      - application/json
      description: Returns the authenticated user's wallet transaction history
      parameters:
      - description: Limit (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.TransactionListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Get my transactions
      tags:
      - wallet
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"